Refactor cell content saving function
