+++ src/main/js/models/Tables.js
@@ -5,220 +5,15 @@ import Dispatcher from "../dispatcher/Dispatcher";
 import {ActionTypes, ColumnKinds} from "../constants/TableauxConstants";
 import ActionCreator from "../actions/ActionCreator";
 import Row from "./Row";
-import {cellModelSavingError, noPermissionAlertWithLanguage} from "../components/overlay/ConfirmationOverlay.jsx";
+import {noPermissionAlertWithLanguage} from "../components/overlay/ConfirmationOverlay.jsx";
 import {
-  canUserChangeCell,
-  getUserCountryCodesAccess,
   getUserLanguageAccess,
   isUserAdmin,
-  reduceValuesToAllowedCountries,
-  reduceValuesToAllowedLanguages
 } from "../helpers/accessManagementHelper";
-import request from "superagent";
 import * as f from "lodash/fp";
-import Raven from "raven-js";
 import i18n from "i18next";
 import React from "react";
-import {remember} from "../components/table/undo/undoer";
-
-const broadcastCellChange = (cell, oldValue, options) => {
-  ActionCreator.broadcastDataChange({
-    cell,
-    row: cell.row
-  });
-  if (!f.matchesProperty("type", "UNDO", options)) {
-    remember(cell, oldValue);
-  }
-};
-
-// sets or removes a *single* link to/from a link cell
-const changeLinkCell = ({cell, value, options}) => {
-  const curValue = cell.value;
-
-  window.devLog("Cell Model: changing cell", cell.id, "value from:", cell.value, "to", value);
-
-  if (f.size(value) > 1
-    && f.size(curValue) === f.size(value)
-    && f.size(f.intersection(value.map(f.get("id")), curValue.map(f.get("id")))) === f.size(curValue)
-  ) { // reordering happened
-    const swappers = f.flow(
-      f.reject(([a, b]) => f.get("id", a) === f.get("id", b)), // remove all elements which kept their position
-      f.map(f.dropRight(1)),                                   // reduce remaining tuples to single entries in two steps
-      f.flatten,                                               //   thus only the swapped elements remain
-      f.map("id"),                                             // retrieve ids of swapped elements in order after swap
-    )(f.zip(value, cell.value)); // Input: tuples of element at index i before and after swap
-    cell.set({value: f.map(f.pick(["id", "value"]), value)}); // don't store LinkOverlay's display value
-    const sortUrl = `/tables/${cell.tableId}/columns/${cell.column.id}/rows/${cell.row.id}/link/${f.first(swappers)}/order`;
-    return new Promise(
-      (resolve, reject) => {
-        if (f.isEmpty(swappers)) {
-          reject("Something went horribly wrong. Swappers:", swappers);
-        }
-        request
-          .put(apiUrl(sortUrl))
-          .send({
-            location: "before",
-            id: swappers[1]
-          })
-          .end(
-            (err, response) => {
-              if (err) {
-                cell.set({value: curValue});
-                reject(err);
-              } else {
-                broadcastCellChange(cell, curValue, options);
-                resolve(response);
-              }
-            }
-          );
-      }
-    );
-  }
-
-  const rowDiff = f.xor(curValue.map(link => link.id), value.map(link => link.id));
-  if (f.size(rowDiff) > 1) { // multiple new values, set new array
-    return new Promise(
-      (resolve, reject) => {
-        cell.save({value}, {
-          success: () => {
-            broadcastCellChange(cell, curValue, options);
-            resolve(cell.value);
-          },
-          error: err => reject("Could not set multiple link values:", err)
-        });
-      }
-    );
-  }
-
-  // Else a single link was added or removed
-  const [toggledRowId] = rowDiff;
-  if (!toggledRowId) {
-    return new Promise((resolve, reject) => {
-      reject("Tried to toggle zero links");
-    });
-  }
-  const {rowId, tableId} = cell;
-  const colId = cell.column.id;
-  const backendUrl = apiUrl(`/tables/${tableId}/columns/${colId}/rows/${rowId}`);
-  cell.set({value: value}); // set locally so fast follow-up request will have correct state
-  const xhrRequest = (curValue.length > value.length)
-    ? request.delete(`${backendUrl}/link/${toggledRowId}`)
-    : request
-      .patch(backendUrl)
-      .send({value: toggledRowId})
-      .set("Content-Type", "application/json");
-  return new Promise(
-    (resolve, reject) => {
-      xhrRequest.end((error, response) => {
-        if (error) {
-          console.warn(error);
-          cell.set({value: curValue}); // rollback local state when anything went wrong
-          cellModelSavingError(error); // this saves us from calculating and undoing diff ourselves
-          reject(error);
-        } else {
-          broadcastCellChange(cell, curValue, options);
-          resolve(value);
-        }
-      });
-    });
-};
-
-export const changeCell = payload => {
-  Raven.captureBreadcrumb({message: "Cell modified", data: {cell: payload.cell.id}});
-  Raven.captureMessage("Cell edited", {level: "info"});
-  if (payload.cell.isLink) {
-    return changeLinkCell(payload);
-  }
-  const {cell} = payload;
-  const actionOptions = payload.options;
-  const oldValue = cell.value;
-  let newValue = payload.value; // value we send to the server
-  let mergedValue; // The value we display for the user
-  let updateNecessary = false;
-  let isPatch = false;
-
-  // Setup for saving the cell
-  if (cell.isMultiLanguage) {
-    mergedValue = f.merge(oldValue, newValue);
-    newValue = {value: newValue};
-    updateNecessary = !f.equals(oldValue, mergedValue);
-    isPatch = true;
-  } else {
-    updateNecessary = !f.equals(oldValue, newValue);
-    mergedValue = newValue;
-    newValue = {value: newValue};
-  }
-
-  return new Promise(
-    (resolve, reject) => {
-      if (updateNecessary) {
-        /**
-         * Basic language access management
-         */
-        if (!isUserAdmin()) {
-          if (!canUserChangeCell(cell)) {
-            noPermissionAlertWithLanguage(getUserLanguageAccess());
-            return;
-          } else {
-            if (cell.isMultiCountry()) {
-              newValue = reduceValuesToAllowedCountries(newValue);
-              if (f.isEmpty(newValue.value)) {
-                // The user tried to change a multilanguage cell without language permission
-                noPermissionAlertWithLanguage(getUserLanguageAccess(), getUserCountryCodesAccess());
-                return;
-              }
-            } else {
-              // reduce values to send just authorized language values to server
-              newValue = reduceValuesToAllowedLanguages(newValue);
-              if (f.isEmpty(newValue.value)) {
-                // The user tried to change a multilanguage cell without language permission
-                noPermissionAlertWithLanguage(getUserLanguageAccess(), getUserCountryCodesAccess());
-                return;
-              }
-            }
-          }
-        }
-        /**
-         * End basic language access management
-         */
-
-        window.devLog("Cell Model: changing cell", cell.id, "value from:", cell.value, "to", newValue.value);
-        // we give direct feedback for user
-        cell.value = mergedValue;
-
-        // we need to clear the newValue, otherwise ampersand save method is merging a strange object
-        if (!isPatch) {
-          newValue = null;
-        }
-
-        /*
-         We want to wait to prevent flashes. We set the value explicitly before saving.
-         Without wait:true save overrides the model for a short time with just one multilanguage value
-         */
-        cell.save(newValue, {
-          patch: isPatch,
-          wait: true,
-          success(model, data, options) {
-            broadcastCellChange(cell, oldValue, actionOptions);
-            // is there new data from the server?
-            if (!f.equals(data.value, mergedValue)) {
-              window.devLog("Cell model saved successfully. Server data changed meanwhile:", data.value, mergedValue);
-              cell.value = data.value;
-            }
-            resolve();
-          },
-          error(error) {
-            cellModelSavingError(error);
-            cell.value = oldValue;
-            reject(error);
-          }
-        });
-      } else {
-        window.devLog("No update required");
-        resolve(newValue);
-      }
-    });
-};
+import changeCell from "./helpers/changeCell";
 
 const Tables = Collection.extend({
   model: Table,
@@ -252,7 +47,8 @@ const Tables = Collection.extend({
   },
 
   changeCellHandler(payload) {
-    const cb = payload.cb || function () {};
+    const cb = payload.cb || function () {
+    };
     changeCell(payload).then(cb);
   },
 
@@ -333,4 +129,5 @@ const Tables = Collection.extend({
 
 });
 
+export {changeCell};
 export default Tables;
